\docType{data}
\name{BlandAltman95}
\alias{BlandAltman95}
\title{Repeated measurements of pH and PaCO2 for eight subjects.}
\format{a data frame with 47 rows and 3 variables}
\usage{
  BlandAltman95
}
\description{
  A dataset used as an example in several papers dealing
  with correlation and panel data.  Contains repeated
  measurements of two variables for eight subjects.  The
  data is unbalanced, ranging from 3--9 observations per
  subject (total 47).
}
\details{
  The variables are: \describe{ \item{patient}{subject
  identifier} \item{pH}{intramural pH measurement}
  \item{paCO2}{intramural paCO2 measurement} }
}
\references{
  Bland, J. M. and Altman, D. G. (1995). Calculating
  correlation coefficients with repeated observations: Part
  1 --- correlation within subjects.  BMJ, 310(446).
}
\keyword{datasets}

